Split CLI into library and binary
