+++ packages/engine/Cargo.toml
@@ -40,7 +40,7 @@ serde-aux = "0.6.1"
 strum_macros = "0.19.4"
 surf = "2.0.0"
 thiserror = "1.0.21"
-tokio = { version = "1.5.0", features = ["macros", "rt-multi-thread", "sync", "process", "io-util", "net", "rt", "fs", "time"] }
+tokio = { version = "1.5.0", features = ["macros", "rt-multi-thread", "sync", "process", "time"] }
 tracing-subscriber = { version = "0.3.3", features = ["env-filter"] }
 tracing-error = "0.2.0"
 uuid = { version = "0.8.1", features = ["v4", "serde"] }
