+++ packages/engine/bin/cli/src/main.rs
@@ -1,23 +1,11 @@
 #![allow(clippy::module_inception)]
 
-#[macro_use]
-extern crate lazy_static;
-#[macro_use]
-extern crate tracing;
-
-pub mod experiment;
-pub mod exsrv;
-pub mod manifest;
-pub mod process;
-
-use std::path::PathBuf;
+use std::{fmt::Debug, path::PathBuf, time::Duration};
 
 use clap::{AppSettings, Parser};
-use error::Result;
-use experiment::run_experiment;
+use error::{report, Result, ResultExt};
 use hash_engine::utils::OutputFormat;
-
-use crate::exsrv::create_server;
+use orchestrator::{create_server, Experiment, ExperimentConfig, Manifest};
 
 /// Arguments passed to the CLI
 #[derive(Debug, Parser)]
@@ -45,13 +33,21 @@ pub struct Args {
     #[clap(long, default_value = "full", arg_enum, env = "HASH_EMIT")]
     emit: OutputFormat,
 
+    /// Engine start timeout in seconds
+    #[clap(long, default_value = "2", env = "ENGINE_START_TIMEOUT")]
+    start_timeout: u64,
+
+    /// Engine wait timeout in seconds
+    #[clap(long, default_value = "60", env = "ENGINE_WAIT_TIMEOUT")]
+    wait_timeout: u64,
+
     /// Experiment type to be run.
     #[clap(subcommand)]
     r#type: ExperimentType,
 
     /// Max number of parallel workers (must be power of 2).
-    #[clap(short = 'w', long, default_value = "4", env = "HASH_WORKERS")]
-    num_workers: u16,
+    #[clap(short = 'w', long, env = "HASH_WORKERS")]
+    num_workers: Option<usize>,
 }
 
 /// Type of experiment to be run.
@@ -64,6 +60,21 @@ pub enum ExperimentType {
     #[clap(name = "simple")]
     SimpleExperiment(SimpleExperimentArgs),
     // Generate shell completitions
+}
+
+impl From<ExperimentType> for orchestrator::ExperimentType {
+    fn from(t: ExperimentType) -> Self {
+        match t {
+            ExperimentType::SimpleExperiment(simple) => orchestrator::ExperimentType::Simple {
+                name: simple.experiment_name,
+            },
+            ExperimentType::SingleRunExperiment(single) => {
+                orchestrator::ExperimentType::SingleRun {
+                    num_steps: single.num_steps,
+                }
+            }
+        }
+    }
 }
 
 /// Single Run Experiment.
@@ -93,11 +104,37 @@ async fn main() -> Result<()> {
             .duration_since(UNIX_EPOCH)
             .unwrap()
             .as_millis();
-        format!("ipc://hash-orchestrator-{}", now)
+        format!("ipc://hash-orchestrator-{now}")
     };
 
-    let (mut experiment_server, handler) = create_server(&nng_listen_url)?;
+    let (mut experiment_server, handler) = create_server(nng_listen_url)?;
     tokio::spawn(async move { experiment_server.run().await });
-    run_experiment(args, handler).await?;
-    Ok(())
+
+    let absolute_project_path = args
+        .project
+        .canonicalize()
+        .wrap_err_lazy(|| format!("Could not canonicalize project path: {:?}", args.project))?;
+    let manifest = Manifest::from_local(&absolute_project_path)
+        .wrap_err_lazy(|| format!("Could not read local project {absolute_project_path:?}"))?;
+    let experiment_run = manifest
+        .read(args.r#type.into())
+        .wrap_err("Could not read manifest")?;
+
+    let experiment = Experiment::new(ExperimentConfig {
+        num_workers: args.num_workers.unwrap_or_else(num_cpus::get),
+        emit: args.emit,
+        output_folder: args.output,
+        engine_start_timeout: Duration::from_secs(args.start_timeout),
+        engine_wait_timeout: Duration::from_secs(args.wait_timeout),
+    });
+
+    let project_name = args.project_name.clone().unwrap_or(
+        absolute_project_path
+            .file_name()
+            .ok_or_else(|| report!("Project path didn't point to a directory: {absolute_project_path:?}"))? // Shouldn't be able to fail as we canonicalize above
+            .to_string_lossy()
+            .to_string(),
+    );
+
+    experiment.run(experiment_run, project_name, handler).await
 }
