+++ packages/engine/lib/orchestrator/Cargo.toml
@@ -1,0 +1,21 @@
+[package]
+name = "orchestrator"
+version = "0.0.0"
+edition = "2021"
+
+[dependencies]
+error = { path = "../../lib/error", features = ["spantrace"] }
+hash_engine = { path = "../..", default-features = false }
+#server = { path = "../server", artifact = "bin" }
+
+async-trait = "0.1.48"
+rand = "0.8.3"
+rand_distr = "0.4.2"
+serde = "1.0.133"
+serde_json = "1.0.74"
+tracing = "0.1.29"
+tokio = { version = "1.5.0" }
+
+[features]
+default = ["build-nng"]
+build-nng = ["hash_engine/build-nng"]
