+++ packages/engine/lib/orchestrator/src/process/local.rs
@@ -81,12 +81,11 @@ impl LocalCommand {
 #[async_trait]
 impl process::Command for LocalCommand {
     async fn run(self: Box<Self>) -> Result<Box<dyn process::Process + Send>> {
-        let env_process_path = std::env::var("ENGINE_PATH");
-        let process_path = if env_process_path.is_err() {
-            PROCESS_PATH_DEFAULT
-        } else {
-            env_process_path.as_ref().unwrap()
-        };
+        let engine_path = std::env::var("ENGINE_PATH");
+        let process_path = engine_path
+            .as_ref()
+            .map(String::as_str)
+            .unwrap_or(PROCESS_PATH_DEFAULT);
 
         let mut cmd = std::process::Command::new(process_path);
         cmd.arg("--experiment-id")
